// frontend/types/types.ts
export interface WalletDeposit {
      id: string;
      user_id: string;
      wallet_id: string;
      wallet_tx_id: string;
      status: 'CONFIRMED' | 'PENDING' | 'FAILED';
      currency: 'NGN'
      amount: number;
      flw_status: string;
      flw_ref: string;
      tx_ref: string;
      payment_type: 'bank_transfer' | 'ussd' | 'account' | 'card'
      transaction_id: number;
      flw_created_at: string;
      created_at: string;

}